+++ Source/Core/Common/IndexPolicies.hpp
@@ -6,17 +6,17 @@ namespace Math
 
 struct IndexDim3Policy
 {
-  real& operator()(Matrix3& A, uint row, uint col)
+  real& operator()(Matrix3& A, size_t row, size_t col)
   {
     return A(row, col);
   }
 
-  real& operator()(Vector3& v, uint i)
+  real& operator()(Vector3& v, size_t i)
   {
     return v[i];
   }
 
-  uint GetDimension(Vector3& v)
+  size_t GetDimension(Vector3& v)
   {
     return 3;
   }
@@ -24,17 +24,17 @@ struct IndexDim3Policy
 
 struct IndexDim4Policy
 {
-  real& operator()(Matrix4& A, uint row, uint col)
+  real& operator()(Matrix4& A, size_t row, size_t col)
   {
     return A(row, col);
   }
 
-  real& operator()(Vector4& v, uint i)
+  real& operator()(Vector4& v, size_t i)
   {
     return v[i];
   }
 
-  uint GetDimension(Vector4& v)
+  size_t GetDimension(Vector4& v)
   {
     return 4;
   }
@@ -43,19 +43,19 @@ struct IndexDim4Policy
 struct GenericDimIndexPolicy
 {
   template <typename MatrixType>
-  real& operator()(MatrixType& A, uint row, uint col)
+  real& operator()(MatrixType& A, size_t row, size_t   col)
   {
     return A(row, col);
   }
 
   template <typename VectorType>
-  real& operator()(VectorType& v, uint i)
+  real& operator()(VectorType& v, size_t i)
   {
     return v[i];
   }
 
   template <typename VectorType>
-  uint GetDimension(VectorType& v)
+  size_t GetDimension(VectorType& v)
   {
     return v.GetSize();
   }
